♻️ Refactor Category View Model to better separation of concerns
Update the Category View Model to be only responsible to load the
categories since it will be reused in several places. The
`updateCategory` was updated to the TaskDetailViewModel, which is the
only place that will update it inside a Task at the moment.
