+++ features/task/src/androidTest/java/com/escodro/task/TaskDetailTest.kt
@@ -5,8 +5,8 @@ import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.test.platform.app.InstrumentationRegistry
 import com.escodro.task.model.Task
-import com.escodro.task.presentation.detail.TaskDetailActions
 import com.escodro.task.presentation.category.CategoryState
+import com.escodro.task.presentation.detail.TaskDetailActions
 import com.escodro.task.presentation.detail.main.TaskDetailRouter
 import com.escodro.task.presentation.detail.main.TaskDetailState
 import com.escodro.theme.AlkaaTheme
