+++ features/task/src/main/java/com/escodro/task/di/TaskModule.kt
@@ -5,8 +5,8 @@ import com.escodro.task.mapper.CategoryMapper
 import com.escodro.task.mapper.TaskMapper
 import com.escodro.task.mapper.TaskWithCategoryMapper
 import com.escodro.task.presentation.add.AddTaskViewModel
+import com.escodro.task.presentation.category.TaskCategoryViewModel
 import com.escodro.task.presentation.detail.alarm.TaskAlarmViewModel
-import com.escodro.task.presentation.detail.category.TaskCategoryViewModel
 import com.escodro.task.presentation.detail.main.TaskDetailViewModel
 import com.escodro.task.presentation.list.TaskListViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
@@ -19,8 +19,8 @@ val taskModule = module {
 
     // Presentation
     viewModel { TaskListViewModel(get(), get(), get()) }
-    viewModel { TaskDetailViewModel(get(), get(), get(), get()) }
-    viewModel { TaskCategoryViewModel(get(), get(), get()) }
+    viewModel { TaskDetailViewModel(get(), get(), get(), get(), get()) }
+    viewModel { TaskCategoryViewModel(get(), get()) }
     viewModel { TaskAlarmViewModel(get(), get(), get(), get()) }
     viewModel { AddTaskViewModel(get()) }
