+++ features/task/src/main/java/com/escodro/task/presentation/category/TaskCategoryViewModel.kt
@@ -1,20 +1,13 @@
-package com.escodro.task.presentation.detail.category
+package com.escodro.task.presentation.category
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import com.escodro.domain.usecase.category.LoadAllCategories
-import com.escodro.domain.usecase.task.UpdateTaskCategory
 import com.escodro.task.mapper.CategoryMapper
-import com.escodro.task.presentation.category.CategoryState
-import com.escodro.task.presentation.detail.main.CategoryId
-import com.escodro.task.presentation.detail.main.TaskId
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.launch
 
 internal class TaskCategoryViewModel(
     private val loadAllCategories: LoadAllCategories,
-    private val updateTaskCategory: UpdateTaskCategory,
     private val categoryMapper: CategoryMapper
 ) : ViewModel() {
 
@@ -27,9 +20,5 @@ internal class TaskCategoryViewModel(
         } else {
             emit(CategoryState.Empty)
         }
-    }
-
-    fun updateCategory(taskId: TaskId, categoryId: CategoryId) = viewModelScope.launch {
-        updateTaskCategory(taskId = taskId.value, categoryId = categoryId.value)
     }
 }
