+++ features/task/src/main/java/com/escodro/task/presentation/detail/main/TaskDetail.kt
@@ -20,13 +20,13 @@ import androidx.compose.ui.tooling.preview.Preview
 import com.escodro.task.R
 import com.escodro.task.model.Category
 import com.escodro.task.model.Task
+import com.escodro.task.presentation.category.CategorySelection
+import com.escodro.task.presentation.category.CategoryState
+import com.escodro.task.presentation.category.TaskCategoryViewModel
 import com.escodro.task.presentation.detail.LeadingIcon
 import com.escodro.task.presentation.detail.TaskDetailActions
 import com.escodro.task.presentation.detail.alarm.AlarmSelection
 import com.escodro.task.presentation.detail.alarm.TaskAlarmViewModel
-import com.escodro.task.presentation.category.CategorySelection
-import com.escodro.task.presentation.category.CategoryState
-import com.escodro.task.presentation.detail.category.TaskCategoryViewModel
 import com.escodro.theme.AlkaaTheme
 import com.escodro.theme.components.DefaultIconTextContent
 import com.escodro.theme.temp.getViewModel
@@ -60,7 +60,7 @@ private fun TaskDetailLoader(
     val taskDetailActions = TaskDetailActions(
         onTitleChanged = { title -> detailViewModel.updateTitle(id, title) },
         onDescriptionChanged = { desc -> detailViewModel.updateDescription(id, desc) },
-        onCategoryChanged = { categoryId -> categoryViewModel.updateCategory(id, categoryId) },
+        onCategoryChanged = { categoryId -> detailViewModel.updateCategory(id, categoryId) },
         onAlarmUpdated = { calendar -> alarmViewModel.updateAlarm(id, calendar) },
         onIntervalSelected = { interval -> alarmViewModel.setRepeating(id, interval) }
     )
